+++ src/components/MyFixedIncomes/index.jsx
@@ -1,11 +1,6 @@
 import { useState, useEffect } from "react";
 
-import {
-  MyFixedIncomesContainer,
-  MyFixedIncomesHeader,
-  MyFixedIncomesBody,
-  MyFixedIncomesRow,
-} from "./styles";
+import { MyFixedIncomesContainer, MyFixedIncomesHeader, MyFixedIncomesBody, MyFixedIncomesRow } from "./styles";
 import { FixedIncomeTitleCard } from "./FixedIncomeTitleCard";
 import { FixedIncomePostionCard } from "./FixedIncomePositionCard";
 import { FixedIncomeDueDateCard } from "./FixedIncomeDueDateCard";
@@ -16,6 +11,7 @@ import SearchIcon from "../../assets/icons/search.svg";
 import { Pagination } from "./Pagination";
 import { validateFilterDate } from "../../utils/validateFilterData";
 import { debounce } from "../../utils/debounce";
+import { usePagination } from "../../hooks/usePagination";
 
 const filterOptions = [
   "Saldo Bruto",
@@ -30,32 +26,13 @@ const filterOptions = [
 
 export function MyFixedIncomes({ data }) {
   const [fixedIncomeData, setFixedIncomeData] = useState([]);
-  console.log(fixedIncomeData);
-
-  const [page, setPage] = useState(1);
-  const [initialPosition, setInitialPosition] = useState(0);
   const [searchTerm, setSearchTerm] = useState("");
 
-  const limitPerPage = 5;
-
-  const pagesQtd = Math.ceil(fixedIncomeData.length / limitPerPage);
-  const dataLength = page * limitPerPage;
-
-  const pagesNumbersForRender = new Array(pagesQtd).fill(0).map((item, index) => index + 1);
+  const { page, initialPosition, dataLength, pagesNumbersForRender, handleChangePagination } = usePagination(fixedIncomeData);
 
   useEffect(() => {
     setFixedIncomeData(data);
   }, [data]);
-
-  function handleChangePagination(pageNumber) {
-    setPage(pageNumber);
-
-    if (pageNumber > page) {
-      setInitialPosition((state) => state + 6);
-    } else {
-      setInitialPosition((state) => state - 6);
-    }
-  }
 
   function handleChange(event) {
     setFixedIncomeData(data);
@@ -65,23 +42,12 @@ export function MyFixedIncomes({ data }) {
     setFixedIncomeData(filteredData);
   }
 
-  function debounce(callback, wait) {
-    let timeout;
-
-    return (...args) => {
-      clearTimeout(timeout);
-      timeout = setTimeout(() => callback(...args), wait);
-    };
-  }
-
   function filterByTerm(fixedIncomesData, data, term) {
     if (searchTerm === "") {
       return data;
     }
 
-    const filteredData = fixedIncomesData.filter((income) =>
-      income.fixedIncome.name.toLowerCase().includes(term.toLowerCase())
-    );
+    const filteredData = fixedIncomesData.filter((income) => income.fixedIncome.name.toLowerCase().includes(term.toLowerCase()));
 
     return filteredData;
   }
@@ -110,26 +76,16 @@ export function MyFixedIncomes({ data }) {
             <div key={income.fixedIncome.portfolioProductId}>
               {index >= initialPosition - 1 && index < dataLength && (
                 <MyFixedIncomesRow>
-                  <FixedIncomeTitleCard
-                    incomeClass={income.fixedIncome.bondType}
-                    description={income.fixedIncome.name}
-                  />
+                  <FixedIncomeTitleCard incomeClass={income.fixedIncome.bondType} description={income.fixedIncome.name} />
                   <FixedIncomePostionCard data={income.position} />
-                  <FixedIncomeDueDateCard
-                    daysUntilExpiration={income.due.daysUntilExpiration}
-                    date={income.due.date}
-                  />
+                  <FixedIncomeDueDateCard daysUntilExpiration={income.due.daysUntilExpiration} date={income.due.date} />
                 </MyFixedIncomesRow>
               )}
             </div>
           );
         })}
       </MyFixedIncomesBody>
-      <Pagination
-        handleChangePage={handleChangePagination}
-        pages={pagesNumbersForRender}
-        actualPage={page}
-      />
+      <Pagination handleChangePage={handleChangePagination} pages={pagesNumbersForRender} actualPage={page} />
     </MyFixedIncomesContainer>
   );
 }
